+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/_internal/PluginIds.kt
@@ -25,7 +25,6 @@ internal object PluginIds {
     object Gradle {
         const val application = "org.gradle.application"
         const val javaGradlePlugin = "java-gradle-plugin"
-        const val kotlinDsl = "org.gradle.kotlin.kotlin-dsl"
         const val versionCatalog = "org.gradle.version-catalog"
     }