+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/gradle/KotlinGradleExtension.kt
@@ -21,7 +21,8 @@ constructor(
 
     @HubdleDslMarker
     public fun Project.plugin(action: Action<HubdleKotlinGradlePluginExtension> = Action {}) {
-        pluginManager.apply(PluginIds.Gradle.kotlinDsl)
+        pluginManager.apply(PluginIds.Gradle.javaGradlePlugin)
+        pluginManager.apply(PluginIds.Kotlin.jvm)
         plugin.run { isEnabled = true }
         action.execute(plugin)
     }